+++ src/bda/plone/ticketshop/at/__init__.py
@@ -1,3 +1,7 @@
+from datetime import datetime
+from DateTime import DateTime
+from DateTime.DateTime import safelocaltime
+from DateTime.interfaces import DateTimeError
 from Products.Archetypes import atapi
 from Products.Archetypes.interfaces import IFieldDefaultProvider
 from Products.Archetypes.public import BooleanField
@@ -9,12 +13,15 @@ from archetypes.schemaextender.interfaces import IExtensionField
 from archetypes.schemaextender.interfaces import IOrderableSchemaExtender
 from bda.plone.orders.interfaces import ITrading
 from bda.plone.shipping.interfaces import IShippingItem
+from bda.plone.shop.interfaces import IBuyablePeriod
 from bda.plone.ticketshop import messageFactory as _
 from bda.plone.ticketshop import permissions
 from bda.plone.ticketshop.interfaces import IBuyableEvent
 from bda.plone.ticketshop.interfaces import IBuyableEventData
 from bda.plone.ticketshop.interfaces import ISharedStock
 from bda.plone.ticketshop.interfaces import ISharedStockData
+from bda.plone.ticketshop.interfaces import ISharedBuyablePeriod
+from bda.plone.ticketshop.interfaces import ISharedBuyablePeriodData
 from bda.plone.ticketshop.interfaces import ITicket
 from bda.plone.ticketshop.interfaces import ITicketOccurrence
 from bda.plone.ticketshop.interfaces import ITicketShopExtensionLayer
@@ -165,24 +172,24 @@ class TicketShippingItem(object):
 
 
 @implementer(IExtensionField)
-class SharedStockExtensionField(object):
+class SharedDataExtensionField(object):
+
+    @property
+    def data_accessor_interface(self):
+        raise NotImplementedError(u"Abstract ``SharedDataExtensionField`` "
+                                  u"does not implement "
+                                  u"``data_accessor_interface``")
 
     def get(self, instance, **kwargs):
         __traceback_info__ = (self.getName(), instance, kwargs)
-        return ISharedStockData(instance).get(self.getName())
+        return self.data_accessor_interface(instance).get(self.getName())
 
     def getRaw(self, instance, **kwargs):
         return self.get(instance, **kwargs)
 
     def set(self, instance, value, **kwargs):
-        if value == '':
-            value = None
-        elif value is not None:
-            __traceback_info__ = (self.getName(), instance, value, kwargs)
-            if isinstance(value, basestring):
-                value = value.replace(',', '.')
-            value = float(value)
-        ISharedStockData(instance).set(self.getName(), value)
+        __traceback_info__ = (self.getName(), instance, value, kwargs)
+        self.data_accessor_interface(instance).set(self.getName(), value)
 
     def getAccessor(self, instance):
         def accessor(**kw):
@@ -211,12 +218,42 @@ class SharedStockExtensionField(object):
             return getattr(instance, name)
 
 
+class SharedStockExtensionField(SharedDataExtensionField):
+
+    @property
+    def data_accessor_interface(self):
+        return ISharedStockData
+
+
 class XSharedStockFloatField(SharedStockExtensionField, FloatField):
-    pass
+
+    def set(self, instance, value, **kwargs):
+        if value == '':
+            value = None
+        elif value is not None:
+            __traceback_info__ = (self.getName(), instance, value, kwargs)
+            if isinstance(value, basestring):
+                value = value.replace(',', '.')
+            value = float(value)
+        super(XSharedStockFloatField, self).set(instance, value, **kwargs)
 
 
 class XSharedStockBooleanField(SharedStockExtensionField, BooleanField):
-    pass
+
+    def get(self, instance, **kwargs):
+        val = super(XSharedStockBooleanField, self).get(instance, **kwargs)
+        return bool(val)
+
+    def getRaw(self, instance, **kwargs):
+        val = super(XSharedStockBooleanField, self).getRaw(instance, **kwargs)
+        return bool(val)
+
+    def set(self, instance, value, **kwargs):
+        if not value or value == '0' or value == 'False':
+            value = False
+        else:
+            value = True
+        super(XSharedStockBooleanField, self).set(instance, value, **kwargs)
 
 
 @implementer(IFieldDefaultProvider)
@@ -252,3 +289,76 @@ class SharedStockExtender(ExtenderBase):
             ),
         ),
     ]
+
+
+class SharedBuyablePeriodExtensionField(SharedDataExtensionField):
+
+    @property
+    def data_accessor_interface(self):
+        return ISharedBuyablePeriodData
+
+
+class XSharedBuyablePeriodDateTimeField(SharedBuyablePeriodExtensionField,
+                                        atapi.DateTimeField):
+
+    def set(self, instance, value, **kwargs):
+        if not value:
+            value = None
+        elif not isinstance(value, DateTime):
+            try:
+                value = DateTime(value)
+                if value.timezoneNaive():
+                    zone = value.localZone(safelocaltime(value.timeTime()))
+                    parts = value.parts()[:-1] + (zone,)
+                    value = DateTime(*parts)
+            except DateTimeError:
+                value = None
+        super(XSharedBuyablePeriodDateTimeField, self).set(
+            instance, value, **kwargs)
+
+
+@adapter(ISharedBuyablePeriod)
+class SharedBuyablePeriodExtender(ExtenderBase):
+    layer = ITicketShopExtensionLayer
+
+    fields = [
+        XSharedBuyablePeriodDateTimeField(
+            name='buyable_effective',
+            schemata='Shop',
+            widget=atapi.CalendarWidget(
+                label=_(u'label_buyable_effective_date',
+                        default=u'Buyable effective date'),
+            ),
+        ),
+        XSharedBuyablePeriodDateTimeField(
+            name='buyable_expires',
+            schemata='Shop',
+            widget=atapi.CalendarWidget(
+                label=_(u'label_buyable_expiration_date',
+                        default=u'Buyable expiration date'),
+            ),
+        ),
+    ]
+
+
+@implementer(IBuyablePeriod)
+class ATSharedBuyablePeriod(object):
+
+    def __init__(self, context):
+        self.context = context
+
+    @property
+    def effective(self):
+        shared_data = ISharedBuyablePeriodData(self.context)
+        effective = shared_data.get('buyable_effective')
+        if effective:
+            effective = datetime.fromtimestamp(effective)
+        return effective
+
+    @property
+    def expires(self):
+        shared_data = ISharedBuyablePeriodData(self.context)
+        expires = shared_data.get('buyable_expires')
+        if expires:
+            expires = datetime.fromtimestamp(expires)
+        return expires